requirements.txt
Faker==25.9.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
python-dateutil==2.9.0.post0
six==1.16.0
allure-pytest==2.13.5
allure-python-commons==2.13.5